move dep tests to dag
