+++ dag_test.go
@@ -61,3 +61,27 @@ func TestDoYaLikeDAGAddsDependencies(t *testing.T) {
 	require.Contains(t, list, base)
 	require.Contains(t, list, template)
 }
+
+func TestDependenciesValidNoError(t *testing.T) {
+	absoluteFolderPath, err := filepath.Abs("./test_fixtures/deps/valid.hcl")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	p := setupParser(t)
+
+	_, err = p.ParseFile(absoluteFolderPath)
+	require.NoError(t, err)
+}
+
+func TestDependenciesInvalidError(t *testing.T) {
+	absoluteFolderPath, err := filepath.Abs("./test_fixtures/deps/invalid.hcl")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	p := setupParser(t)
+
+	_, err = p.ParseFile(absoluteFolderPath)
+	require.Error(t, err)
+}
